utils: Move link_distro() to distro.py
